фикс расписаний

+++ Control/Scheduler.h
@@ -54,8 +54,8 @@ public:
 	const char* get_note(void); 						// Получить описание
 	const char* get_name(void) { return  "Scheduler"; } // Получить имя
 	int8_t   save(void);  			               		// Записать настройки в eeprom i2c
-	int8_t   load(uint8_t *data = NULL);  		     	// Считать настройки из eeprom i2c
-	int32_t  loadFromBuf(int32_t addr, byte *buf);  	// Считать настройки из буфера на входе адрес с какого, на выходе конечный адрес, число меньше 0 это код ошибки
+	int16_t  load(uint8_t *data = NULL);  		     	// Считать настройки из eeprom i2c, возращает длину или ошибку
+	int8_t   loadFromBuf(int32_t addr, byte *buf);  	// Считать настройки из буфера на входе адрес с какого, код ошибки
 	uint16_t get_crc16(uint8_t *data);             		// Рассчитать контрольную сумму для данных на входе входная сумма на выходе новая
 	int8_t   check_crc16_eeprom(void);
 	uint16_t get_data_size(void) { return sizeof(sch_data); }
